+++ workspace/setup.sh
@@ -1,24 +1,32 @@
 #!/bin/bash
+set -e
 ###################
 #    variables    #
 ###################
 current_dir="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 dir="/home/dorna/Downloads/workspace"
 repo="https://github.com/dorna-robotics/workspace"
+branch="main"
 
 ########################
 #    clone or pull     #
 ########################
-if [ -d "$dir/.git" ]; then
-    cd $dir
-    git fetch origin
-    git reset --hard origin/main
-    git clean -fd
-else
-    rm -rf $dir
-    git clone $repo $dir
-    cd $dir
-fi
+sync_repo() {
+    if [ -d "$dir/.git" ]; then
+        cd "$dir"
+        if git fetch origin "$branch" \
+            && git checkout -B "$branch" "origin/$branch" \
+            && git reset --hard "origin/$branch"; then
+            git clean -fd
+            return 0
+        fi
+        cd /
+    fi
+    rm -rf "$dir"
+    git clone -b "$branch" "$repo" "$dir"
+    cd "$dir"
+}
+sync_repo
 
 # navigate to package and install
 cd workspace
